+++ src/components/pages/NotificationCenter/NotificationItem.tsx
@@ -104,7 +104,7 @@ const NotificationContent = ({
       <div>
         <Trans
           ns="notification"
-          i18nKey={`${item.typeId}.content`}
+          i18nKey={`item.${item.typeId}.content`}
           values={{
             you,
             app: appName,
@@ -269,6 +269,7 @@ export default function NotificationItem({
         page: PAGE,
         size: PAGE_SIZE,
         args: {
+          notificationTypeIds: [],
           notificationTopic: NOTIFICATION_TOPIC.ALL,
           sorting: SORT_OPTION,
         },
@@ -341,7 +342,7 @@ export default function NotificationItem({
               }}
             >
               {' '}
-              {t(`${item.typeId}.title`, {
+              {t(`item.${item.typeId}.title`, {
                 app: item.contentParsed?.AppName ?? item.contentParsed?.appName,
                 offer: item.contentParsed?.OfferName,
                 credentialType: item.contentParsed?.type,